Move `PoolerConfig` from `config/__init__.py` to `config/pooler.py`
